scmstore: don't always flush FileStore on Drop
Summary:
Currently, the scmstore backingstore backend for EdenFs performs very poorly compared to ContentStore. This is mainly because a local-only FileStore is created on every fetch, and always flushes when dropped (even when not used).

With this change, we'll only flush when the parent FileStore is dropped.

This is a slightly hacky fix, vs. creating a new non-flush-on-drop type like `TreeStoreFetch` as we did for trees in the previous diff, but this should work fine for now and is radically smaller. Later I'd like to clean up both to eliminate the verbosity in the trees approach while still using separate types rather than a bool.

Reviewed By: andll

Differential Revision: D31591276

fbshipit-source-id: 11266c19ac68d87015719f5bcbd8f857e596bfdb