test-edenapi-files: switch to debugapi
Summary:
See previous diffs for context. This makes the test shorter and will eventually
allow us to remove some bloated code in edenapi.

Note some hashes are serialized (suboptimally) as an array of bytes, not as a
byte slice.  Ideally we need `serde_bytes` annotation somewhere but I haven't
checked if that breaks compatibility.

Reviewed By: yancouto

Differential Revision: D31465828

fbshipit-source-id: 7892ececc475bad530708499cf2255852d610bc2