types: add an abstract hash type
Summary:
There are a lot of hash types: HgId, Sha256, and edenapi-types has some.
The HgId seems to have most features but it is hard to reuse it for similar
but different types. This is an attempt to do so and unify the main implementation
of those types.

Most methods are copied from HgId. `from_byte_array` is made const fn so it
can be used to construct constants.

The error type is intentionally chosen to be not `anyhow::Error`.
Using static typed errors is considered good practice for low-level crates.
The benefit is that higher level users get their choice - precise static error
type with compile-time checks, or convenient, dynamic error types by anyhow
with runtime downcasts.

Reviewed By: yancouto

Differential Revision: D31615719

fbshipit-source-id: 337356721354c43fe23b9f2d0e90d104c8864c44