types: migrate Sha256 to AbstractHashType
Summary:
See the previous diff for context.

Aside from error type changes, this changes the serialization format of Sha256
from a tuple of 32 u8 to a byte slice. It seems okay, because the only user
of the `Sha256` serialization on client-side is revisionstore's `FileAuxData`,
which does not serialize to disk (This is checked by dropping the Serialize
trait and see what breaks).

Reviewed By: yancouto

Differential Revision: D31615716

fbshipit-source-id: 3d31b5d356c7e5f6b229fa9eae71ba4cad1c0e1a