clarify ImmediateFuture's callback rules and introduce isReady
Summary:
In a meeting today, the question arose "when should one use Future,
SemiFuture, or ImmediateFuture?" To help clarify the situation,
reframe some of ImmediateFuture's doc comments in terms of their
semantics.

Simultaneously, replace hasImmediate with isReady, which returns true
when a value can be retrieved without blocking.

Also, allow converting a completed SemiFuture into an ImmediateFuture
which is also isReady.

Reviewed By: xavierd

Differential Revision: D31563622

fbshipit-source-id: b942031c15d46948b1f60ade9f8043fa19442b65