runlog: automatically clean up old entries
Summary:
When initializing the runlog logger, we probabilistically invoke a cleanup function to delete old log entries. Cleanup happens with a chance of runlog.cleanup_chance [0-1) (i.e. cleanup can be disabled with 0, or run every time with 1). Default chance is 0.01 or 1% of command invocations.

Cleanup iterates over each runlog file and removes it if both:
1. Can acquire shared file lock, meaning the command is no longer running, AND
2. The file's mtime is at least as old as runlog.cleanup_threshold seconds.

Reviewed By: quark-zju

Differential Revision: D31128863

fbshipit-source-id: 7cbb709ab1006e5c86aa265db4ad430db7e69741