Make sure parent is backed up
Summary:
To create a snapshot, we need mononoke to know about the parent commit, that is, it needs to either be public, or uploaded with commit cloud.

Currently, if you create a commit, and very quickly try to create a snapshot on top of it, it will crash because the commit is not still backed up.

This diff makes sure the parent is always backed up before creating the snapshot. In most cases this should be quick and upload nothing.

Reviewed By: StanislavGlebik

Differential Revision: D31610970

fbshipit-source-id: e772711532e226972a90cd53b51a450df4882dde