Wrap gitlfs semaphore in task spawn
Summary:
As described in T102183795, we need to wrap semaphores in task::spawn in order to reduce the risk of deadlocks in case the caller is using a buffered stream.

This change adds task::spawn, and encapsulates the stream in an interface similar to GitPool::with in order to make it more foolproof. This should also let us remove the StreamWithObject hack later.

Note: The code actually locks the semaphore twice with this change. I'll fix that in D31632288, but I wanted to split up the changes to make them easier to review.

Reviewed By: StanislavGlebik

Differential Revision: D31630703

fbshipit-source-id: 7bff0a45022d167ec64d0f169dbb12d5b01cf021