utils: speedup ImmediateFuture::thenValue exception case by 78x
Summary:
Exception cases should be rare, but EdenFS has a couple of spots where
exceptions are somewhat expected. One of which being a lookup failure due to a
non-existing file. Due to the overhead of these lookup failures, EdenFS is
actually spending as much time resolving these, than resolving successful
lookups.

The reason for the high overhead is due to the ImmediateFuture::thenValue just
relying on an exception being thrown when accessing the stored Try value. We
can do better than this by testing the Try before hand, and bailing out in that
case.

Reviewed By: fanzeyi

Differential Revision: D31655223

fbshipit-source-id: 29da7ef38a19f09c91c931e7e8e080e5ed6af269