mononoke: make it possible to derive many types in benchmark mode
Summary:
Previously benchmark mode worked only with a single derived data type. This
diff allows to derive a few types at once using derive graph functionality.

A few things that I'd like to note:
1) While doing so I decided to remove BenchmarkResult that we used before. It
turned out to be not very useful, and it made it harder to add support for
multiple derived data types. For now I decided to remove it, and we can add it
later if needed.
2) Previously batch size could only be set for parallel derivaton mode, however it
makes sense to set it for backfill mode as well. Given that it's useful for 2
derivation type out of 3 I decided to allow to set batch size everywhere.

Reviewed By: ahornby

Differential Revision: D31603912

fbshipit-source-id: 1c0ab2f5ab4bf5a72e3b0a9b99cd91c8c5d9b6d8