mononoke: support gaps in new skeleton_manifest derivation
Summary:
In previous diff we've added support for skeleton manifest derivation that uses
`derive_manifests_for_simple_stack_of_commits` function, but it didn't support
derivation with gaps. This diff adds support for that.

Brief reminder - deriving with gaps is possible for fsnodes and skeleton
manifests, because these manfiests do not depend on history. So for a linear
stack of commits like this

```
A  <- top commit of the stack
|
B
|
C
```

We can skip deriving data for B and C and derive data directly for A. To do
that we need to combine file changes for A, B and C together and then run
derivation as if all these file changes were made by commit A.

Reviewed By: ahornby

Differential Revision: D31681924

fbshipit-source-id: 9a7d2e5e07a47159017c2290994a4aedbdbb1f58