bookmarks_movement: permit deletion of scratch bookmarks
Summary:
In the past we prohibited the deletion of scratch bookmarks as it would
be impossible to sync this over to Mercurial, which doesn't allow for
the deletion of bookmarks from infinitepush.

Now that we are entirely on Mononoke, we can permit these deletions.  This
requires adding a new `delete_scratch` method to bookmark transactions, which
functions the same as other deletions, except they are not logged to the
bookmark update log (as is true for all scratch bookmarks).

Reviewed By: ahornby

Differential Revision: D31505793

fbshipit-source-id: bf41926b698ad566de0be8123d27bb35a483ade1