Make number of queue workers configurable
Summary:
The number of worker threads handling downloading content from source control servers is currently hard-coded. We want to make that value configurable.

This diff removes all hardcoded values and adds a numBackingstoreThreads option to the edenfs config. This value defaults to the original hard-coded value, 32. The minimum number of threads is 1.

As a result, we remove the numberThreads parameter from the HgQueuedBackingStore constructor and modify all code that previously used that parameter.

Reviewed By: kmancini

Differential Revision: D31739881

fbshipit-source-id: fc1706a9a4b7b7b4f4b5d87ac428ddc025616d40