rename ObjectId::toString to toLogString
Summary:
ObjectId::toString was too ambiguous of a name which makes it hard to
reason about the correctness of all of its uses as we generalize the
meaning of object identifiers.

Rename toString to toLogString, and change a bunch of call sites to
use the iostream, folly::to, or fmt formatters.

For the one place that knows the format of the ID but wants its hex
representation, use asHexString. This gives us the freedom to change
the log representation of ObjectIds in the future without worrying
about breakage.

Reviewed By: andll

Differential Revision: D31776897

fbshipit-source-id: 8160ab581f260b3bb955afbad65d6dbe8486e3bc