Allow building ChangesetContext for commit in a bubble
Summary:
This will be used for commits in bubbles, like snapshots. This stack makes SCS be able to query snapshot data as well as usual commits.

This diff also makes SCS start using the bubble id if it's provided in the thrift call, and create the repo changeset using that bubble. This is the "server side" change, the "client side" is D31688753.

Reviewed By: StanislavGlebik

Differential Revision: D31687574

fbshipit-source-id: 1d9be59b8bd08c97f32add2e56f8b56402ae76cc