Add fallback derived data manager
Summary:
This makes it so each DerivedDataManager can have a fallback manager, and a way to decide which manager should derive derived data for given changesets.

This will be used for deriving data for snapshots (and in the future, commits) inside ephemeral bubbles.

This diff also adds a function that creates a fallback derived data manager for deriving data in bubbles.

Reviewed By: StanislavGlebik

Differential Revision: D31400441

fbshipit-source-id: 0c551629059ebbba68700dfce3c194cf47a42979