Filenodes and BonsaiHgMapping should be optional
Summary:
We have managers for bubbles, and they don't have a filenode/bonsai hg mapping equivalent that is ephemeral, so to be on the safe side it's better to make these fields optional.

(This shows that ideally this would structured a bit better, Filenodes is only used by filenodes derivation but it still needs to be on the "main" manager. We couldn't think of a better simple way to do this and that's why we went with this.)

Reviewed By: StanislavGlebik

Differential Revision: D31658650

fbshipit-source-id: caa3b5767bf50a94bc2611f5db87605f103b4535