Forward compatibility with 2021 edition $:pat
Summary:
The meaning of `$:pat` is different from 2015 and 2018 vs the 2021 edition. The old meaning of `$:pat` is now named `$:pat_param`. But `$:pat_param` is also available on old editions and means the same thing as `$:pat` there, so we can adopt it in anticipation of flipping the default edition switch in fbcode.

Without this diff, this macro does not compile in 2021 edition because 2021's `$:pat` is not allowed to be followed by a `|` token.

Details: https://doc.rust-lang.org/edition-guide/rust-2021/or-patterns-macro-rules.html

Reviewed By: StanislavGlebik

Differential Revision: D31823201

fbshipit-source-id: dbb790b3e90148b0b672265f88becec99ba242ff