cli_rs: Strip environment variables of `PYTHONHOME` and `PYTHONPATH`
Summary: Since we have a rust wrapper around the python version of edenfsctl (as we are working on moving to be completely rust), we can strip these environment variables from the fallback python edenfsctl command. Without stripping them, this breaks edenfsctl for some users who have custom PYTHONHOME and PYTHONPATH variables that don't contain the modules the python fallback needs to import.

Reviewed By: fanzeyi

Differential Revision: D31810981

fbshipit-source-id: d357254d7b122d4c501470e1701b034883fca174