cli_rs: Scaffolding for minitop subcommand
Summary:
As a first step, as I ramp up on Rust and the edenfsctl codebase and in order to more quickly support basic top functionality in windows, we're adding a minitop subcommand that will have a subset of the `eden top` output.

This diff adds scaffolding for the minitop subcommand.

Reviewed By: fanzeyi

Differential Revision: D31841478

fbshipit-source-id: 7020e81a0c8021bad45de835f2e279c049b76200