immediatefuture: add collectAll that takes a parameter pack
Summary:
In some places, EdenFS uses folly::collectAll passing futures to it. In order
to replace it, we need an equivalent for ImmediateFuture.

The implementation relies on the fact that ImmediateFuture's constructor will
extract the value from a SemiFuture if it is ready, and thus the code doesn't
have to special case when all the passed in ImmediateFuture are immediate
values.

Reviewed By: chadaustin

Differential Revision: D31667897

fbshipit-source-id: ea680e1c74528d1eb34210cacb58efda8a26fee0