blame_v2: split batches on copy_info
Summary:
When splitting batches into linear stacks for parallel blame derivation, we
must also split on the existence of `copy_info`.

Blame derivation follows copy information, so if a commit in a batch modifies a
file, then a subsequent commit in the batch that copies from that file must be
placed in a separate stack so that it can access the blame data for the
copied-from file.

In theory it would be sufficient to check that copied-from sources aren't
modified by the previous commits in the stack, but this check isn't available
right now, so we fall back to just splitting on any copy info.  We can add more
detailed splitting later if we need to improve performance further.

Reviewed By: StanislavGlebik

Differential Revision: D31898499

fbshipit-source-id: 85edb197231095de5ac17e07d5fab919277614ed