derived_data: log batch derivations to scuba
Summary:
Add scuba logging for batch derivations.

In the request context scuba log, we log the batch start and end.

In the derived data scuba log, we log the full set of changeset ids for
the batch, however we do not log the serialized form of the mapped changeset
(this can be added, but it's likely to slow things down).

Reviewed By: StanislavGlebik

Differential Revision: D31902279

fbshipit-source-id: fda2bd608159a3d5e4c732a1cf76a02c0f7932cb