prjfs: kill timeout notification
Summary:
The timeout based notification mechanism appears to be at the root of a lot of
EdenFS crashes. It also doesn't detect all of the network failures and is being
replaced with a more robust mechanism which will handle these cases.

For now, let's just get rid of the timeout based mechanism.

Reviewed By: fanzeyi

Differential Revision: D31739683

fbshipit-source-id: d6baa03fe3579dc36c9f435bd16df39799e51ef9