debug: add debugdiffdirs for watchman
Summary:
Adds a command which prints out which directories were modified between
two diffs. "modified" is defined as "has an immediate child that has been added
or removed". This means just modifying a file will not cause it's parent
directory to be marked changed.

This is used for watchman to report what directories have changed during an eden
checkout.

Reviewed By: quark-zju

Differential Revision: D31698688

fbshipit-source-id: ef90573af3027306501673821ee40b691241b6e4