service: no longer make edenfs a suid root binary
Summary:
In order to execute the edenfs_privhelper, EdenFS needs to be a suid root
binary, but due to the large amount of dependency, it's hard to fully control
what runs before main and thus what executes with root privileges.

To reduce the attack surface, let's only make the edenfs_privhelper a suid root
binary. In tests, we still start EdenFS with sudo, and thus the
edenfs_privhelper will have the proper privileges.

Reviewed By: chadaustin

Differential Revision: D31870880

fbshipit-source-id: 3f06d22ecd4f3bc65e3e955c5a1de811b1cf328f