indexedlog: add "fold function" concept to Log
Summary:
Prevously, a `Log` has a few index functions. The indexes are useful for
indexing individual items. This diff defines a new concept: "fold function".
It will be useful for accumulated information, such as the total count of
a special kind of entries.

This diff just defines the concept. Implementation will be done later.

Practically, this feature will be used to track the "assigned" segments in
the `dag` crate.

This feature cannot be easily built outside indexedlog because indexedlog
does not provide a way to iterate entries from an arbitrary offset. indexedlog
does not leak the raw offsets either. That makes incremental fold hard
to implement outside indexedlog.

Reviewed By: DurhamG

Differential Revision: D31734838

fbshipit-source-id: d48dca0da1c753d23cee725dc22a809d529c55db