discovery: drop unsound "all remote known" fast path (BC)
Summary:
This used to be correct when both local and remote list all heads.  With
selectivepull, we only ask remote about certain heads, if those heads are
known, that does not mean the common heads are just those, since the server
might have other branches refer to related commits.

Dropping the path allows us to figure out "missing" commits more accurately.
This can be seen from the change in `test-fb-hgext-treemanifest-treeonly.t`.

This change also allows us to reproduce a pushrebase problem in the next diff.

This is a breaking change in case the push does not create new commits while
creating a remote bookmark. Previously the exit code might be 1. Now it's 0,
which seems more desirable since it's not "nothing changed".

Reviewed By: DurhamG

Differential Revision: D31877853

fbshipit-source-id: d03c1d0ffec98b46fbc59ff12ab2015f3c10385c