prjfs: use a FutureSplitter in FileMetadata
Summary:
Futures can only be waited on once, but this code can potentially try to wait
on it more. If for instance the size future were to fail, the continuation
would never be called, and thus the next call to getSize would obtain an
already moved future.

For now, let's simply use a FutureSplitter to close this bug. In the future,
the readdir code will be rewritten to be more asynchronous and to avoid this.

Reviewed By: genevievehelsel

Differential Revision: D31782917

fbshipit-source-id: 9e78fdec62dbccbc2796722175936d563a20ebcd