edenapi: compress file responses in parallel
Summary:
We're seeing ever growing memory usage during large checkouts. It
appears to be caused by lz4 compression occupying the main thread enough that we
don't pull data out of edenapi fast enough, so the data from the server starts
queueing in memory.

Let's have multiple threads compressing the responses.

Note this only fixes it for non-scmstore. scmstore needs this same change, but
testing shows it still has a memory leak after this change, so it's hard to tell
if this change is working for scmstore. I'll do a followup diff for that.

Reviewed By: quark-zju

Differential Revision: D31973093

fbshipit-source-id: a6d2474a98b410085ce531db28a116c87d9e210c