inodes: add a loadAndRemoveOverlayDir function
Summary:
As the name implies, it loads an overlay directory and removes it at the same
time. This is done to bypass calling removeOverlayData when recursively
removing a directory as removeOverlayData would check if the directory no
longer has childrens and thus would fail.

Credits for this diff should go to chadaustin.

Reviewed By: genevievehelsel

Differential Revision: D30843607

fbshipit-source-id: 370491e454990ca5c8215501cd7e1d5e85526cce