mononoke: add --no-upload-if-less-than-chunks option to streaming_changelog
Summary:
We usually had option called "--skip-last-chunk" set, however turned out it's a
dangerous option to use. Normally a streaming changelog has only a single head,
however if we add --skip-last-chunk then there might be multiple heads in the
repo if the skipped chunk contained merge commits.
And this code https://fburl.com/code/0fkgmnaj just uses a "tip" commit to do a
pull right after streaming clone. The problem is that "tip" can be only a
single head out of all the heads we have. And tip can be a commit with a very
low generation number, in that case the subsequent pull would try to pull too
many commits at once.

In order to fix the issue I suggest to add --no-upload-if-less-than-chunks
option. This option wouldn't even try uploading chunks if we have too few of
them. So on the one hand it would allow us to not create too many small chunks,
but on the other it would fix the issues mentioned above.

Note that this option would mean that we would create more chunks than we
otherwise would. For example if --no-upload-if-less-than-chunks is 2, then on
average we would create 2x more chunks.

Reviewed By: Croohand

Differential Revision: D31957177

fbshipit-source-id: 8fa9476e4236d962d5f4f0fae96ce6db996a22dc