Make use of mononoke, mercurial and git thrift types for conversion into DerivedData
Summary:
I've noticed that we have all mentioned thrift types already, with implemented from/into thrift methods. It will come handy in the next diff, for the conversion from DerivedData into different BonsaiDerivable types.
In addition I've renamed the `from` function in DireveConvert trait to not confuse it with `std::convert`.
And while on that I noticed missing git `TreeHandle` handling on the service side.

Reviewed By: StanislavGlebik

Differential Revision: D31888726

fbshipit-source-id: a3fb9e58df4691e0b2cc8b66e007e4e96d3f9f04