edenapi: compress file responses in parallel for scmstore
Summary:
We're seeing ever growing memory usage during large checkouts. It
appears to be caused by lz4 compression occupying the main thread enough that we
don't pull data out of edenapi fast enough, so the data from the server starts
queueing in memory.

Let's have multiple threads compresing the responses.

In theory this should've reduced memory usage, but in reality there's an issue
with the evict cache logic that prevents this. This change is still useful
though, since there was also a bug where us calling stats.await caused 1) the
whole thing to hang waiting for the request to finish before we even processed a
single entry, and 2) waiting on stats first caused a deadlock with the next diff
since stats aren't produced until the first entry is processed.

Reviewed By: quark-zju

Differential Revision: D32013158

fbshipit-source-id: 9b1b79676e12d6185866b436f5c4c2cfd5c94503