patch: Accept contexts and not nodes in patch functions
Summary:
The `edenscm.mercurial.patch.diffhunks`, which is the heart of functionalities like `hg diff` and `hg show`, takes two nodes, and then gets the `changectx` from them to compare the changes.

That works for most cases, but doesn't allow things like `memctx` to be used, as it doesn't have a node.

This refactors so that function (and functions that call it) take a `ctx` instead.

 ---

EDIT: I just found out `repo[ctx]` returns `ctx`, so I could just pass in context as if it was a node (that is, only fix `changeset_printer` to pass `ctx` to `diffordiffstat`), but the fix on this diff still seems better, more consistent and less unclear, so I still think it should be merged.

Reviewed By: DurhamG

Differential Revision: D31802139

fbshipit-source-id: bc9d23a5e666cd0f57b69bce3ed9d58ae79e92d0