Take bubble into account in RepoContext
Summary:
Previously I had added the "bubble override" on `ChangesetContext` level. However, this made snapshots very hard to work with, as you had to be very careful not to let it "slip" back to `RepoContext` and use the non-bubble blobstore/changesets.
To make scsc work in that way, I would have to refactor *a lot* of code to use `ChangesetContext`, and it would be very easy for a bug to be introduced.

This diff makes the override on `RepoContext` level, by changing `BlobRepo` itself, so it's harder to make a mistake and almost everything "just works", like accessing the blobstore or deriving something.

This is still not super pretty, I dislike editing BlobRepo like this, but I don't see a better alternative, other solutions seem more bug-prone.

Reviewed By: markbt

Differential Revision: D31763334

fbshipit-source-id: 78060666bb447a2e854f2010fc55c63ddb6db9aa