check for `--privhelper_path` when deciding whether to use `sudo`
Summary:
When running the Eden tests using buck2, this flag is used because the
privhelper is not immediately next to the EdenFS binary.

This diff updates the approximation of the privhelper path introduced in
D31870880 (https://github.com/facebookexperimental/eden/commit/710f03501d5f5046f55a11308879323fc79156a0) accordingly so that it also works in this case.

Reviewed By: xavierd

Differential Revision: D32070038

fbshipit-source-id: c111cc533bb9e9d111bfe9bad71e817655b3d613