Introduce capabilities instead of version numbers for takeover
Summary:
This commit is introducing capabilities for the takeover code. This will make it
easier to evolve the protocol. Which I am about to do to support NFS.

Why is capabilities worth adding?
Currently takeover uses a versioning based approach to evolve the takeover
protocol. This works and is pretty simple. The downside is that it leads us to
making lots of checks in the code base like: if this version then do x.

There are two downsides of such checks:
1. reading the code is not self contained. You have to know or have already
read which version introduces which features for the checks to make sense. This
generally makes the code harder to read and harder to check in code review.

2. introducing new versions sometimes mean we need to update those checks we
previously added. For example we make this check in the server on wether to
send a ping message before sending data:
```
  shouldPing_ =
      (protocolVersion_ == TakeoverData::kTakeoverProtocolVersionFour);
```

if we add a kTakeoverProtocolVersionFive where we still want to send the ping (
which we likely do because the ping is a part of the protocol correctness under
failures), then we need to update this time to

```
  shouldPing_ =
      (protocolVersion_ >= TakeoverData::kTakeoverProtocolVersionFour);
```

then if we replace the ping with something else we would have to again update
the code to
```
  shouldPing_ =
      (protocolVersion_ >= TakeoverData::kTakeoverProtocolVersionFour
      && protocolVersion_ <= lastTakeoverVersionWithPing);
```

The first of these code updates is very easy to skip accidentally if you are
making an update to the protocol that has nothing to do with ping. This would
silently introduce a regression.

Using capabilities instead makes this check
```
  shouldPing_ =
      (protocolVersionCapabilities_ & TakeoverCapabilities::PING);
```
This is more readable. and this code never has to be updated; we just don't use
the ping capability any more.

Eventually, if we go down the capabilities route we will want to kill off
versions completely. This will mean bumping the version of the protocol, where
we agree on capabilities after agreeing to that version. After this version is
stable (released to production and a good track record indicating we don't need
to rollback prior to the change), we can delete the older versions of the
protocol and all the version based stuff in the code. This stabilizing process
will take a while, so I do not want this to block the NFS evolution I am doing
in the next change. I will comeback to this after we get NFS support in graceful
restart.

Reviewed By: xavierd

Differential Revision: D31965080

fbshipit-source-id: f8413d07da42a92125d46aa3cfbd3fa458b87b46