privhelper: sanity check every mount points
Summary:
The privhelper can be tricked to mount or umount in places of the FS that the
user doesn't actually have access. This can be misused to mount or unmount to
privileged paths.

When accepting mount points from EdenFS, let's verify that the user actually
have access to them first.

Reviewed By: chadaustin

Differential Revision: D31875209

fbshipit-source-id: 939bf8f519adb63ce54f91c7b0757a7486cabf51