Lookup client only fix
Summary:
We recently had to disable edenapi uploads, because it was failing on big commits.

## The issue is
We batch some types of edenapi requests. E.g. if we're trying to check for the existence of 80k files in the blobstore, we split this in two requests, one with 50k ids, and the other with the remaining 30k.
The problem is the client currently depends on the request being done in a single go. For lookup, the client sends a list of ids, and expects responses of type `LookupResponde {index, Option<token>}`, with indexes of each element, and the token in case it was already present.
This does not work when the request is batched: the second request will return invalid indexes when compared with the full array.

This affects **all code** that uses lookup. I believe it is being used for [commit lookup](https://fb.workplace.com/groups/3112141815538647/posts/4178113148941503), but we probably don't reach the problem because it must be rare to lookup thousands of commits at the same time.

## Possible solutions
1. Merge responses smartly: we can modify second responde to add an offset to every index.
2. Don't depend on how the request is batched.

This diff starts the fix to go with solution (1).

## What this diff does
Changes the response from
```
LookupResponse { index, Option<token> }
```
to
```
enum LookupResult {
   Present(token),
   NotPresent(id)
}
LookupResponse { result: LookupResult }
```

That is, we now no longer return indexes, but instead a result enum, that either has the token if blob was uploaded, or the id it was not (so client can tell which blobs are for sure not uploaded, in the future we can maybe just not return those, but it happened currently, so we can't change that)

In practice, we still need to keep all the fields for now, for the transition period.

### Transition:
1. Client supports both old and new server responses. It notices which response it was and converts all of them to the new format. Server response is unaffected. **(this diff)**
2. On server, return both new and old response. Can be landed together with this diff, since we're returning both responses.
3. Delete code that handles old format. Should be landed once (2) is rolled out.

This could go quicker (backwards incompatible change) only if we do not use this in prod, which is not the case.

Reviewed By: StanislavGlebik

Differential Revision: D31935606

fbshipit-source-id: 724098258681630e6941234befd694e6f1ef9e16