auto_wire with enums with transparent variants
Summary:
This allows using the `#[auto_wire]` proc macro on enums with variants that have a single field (which I'm calling "transparent" in the title). This allows covering most of the enums.

Allowing more fields (or struct variants) is possible, but brings some more complications around compatibility, so I avoided this for now.

Reviewed By: StanislavGlebik

Differential Revision: D31935604

fbshipit-source-id: c7e842dbb0b344902905b79997f193096f8ec9e0