mononoke: rename update_wc_equivalence and replace a few usages of it
Summary:
This is the next step in removing current_version (see D32018280 for
motivation).
`update_wc_equivalence` uses current_version, and in most cases it sets the
target commit to None. For this cases let's use set_no_sync_candidate().

For other cases of update_wc_equivalence there was a bug where we use `current_version` instead of the
version that was used to remap commit. Luckily these cases were rare and this
bug never affected us. This diff fixes this issue nevertheless.

Reviewed By: mitrandir77

Differential Revision: D32023698

fbshipit-source-id: 47e9fe2c3042213996bf04c1351862c077d1c3e5